Switch to float16 quantized model from HF Hub

+++ src/main.js
@@ -49,7 +49,7 @@ async function loadModel() {
       'itzune/txukun-cap-punct-eu',  // HF Hub model
       {
         device: 'wasm',
-        dtype: 'fp32',      // use full precision (no _quantized suffix)
+        dtype: 'fp16',      // float16 quantized model (50% smaller)
         subfolder: '',       // model files are directly in repo root, not in onnx/
         progress_callback: (info) => {
           if (info.status === 'progress' && info.progress !== undefined) {
